+++ packages/components-js/CHANGELOG.md
@@ -7,12 +7,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- `Textarea` component
+- `Select Wrapper`
+
+### Fixed
+- `Text Field Wrapper` toggle password visibility
 
 ## [1.0.3] (2020-02-13)
 
 ### Fixed
 - JS framework compatibility
+
+## [1.1.0-0] (2020-02-06)
+
+### Added
+- `Text Field Wrapper`
+
+### Changed
+- Add proper cursor for disabled state for `Button` and `Button Pure`
 
 ## [1.0.2] (2020-02-04)